Add files via upload
Prediction whether a customer is going to churn based on the usage and other demographic factors. As per our data current churn percentage is 17%.

Many Business Development models are present in the market, some use humans as their tool, some use calculations on data as the tool etc. 

The proposed project involves calculations on data and Machine Learning models that predict the results to the problem of growth of company. 

A meaningful insight related to the company’s transaction with the customers. Developing good relations with customer results in higher retention rate and leads to development of company’s customer base.